Shut up Clang about casting from 'unsinged' to pointer

+++ liveMedia/H263plusVideoRTPSink.cpp
@@ -64,7 +64,7 @@ void H263plusVideoRTPSink
     }
     if (frameStart[0] != 0 || frameStart[1] != 0) {
       envir() << "H263plusVideoRTPSink::doSpecialFrameHandling(): unexpected non-zero first two bytes: "
-	      << (void*)(frameStart[0]) << "," << (void*)(frameStart[1]) << "\n";
+	      << (void*)(unsigned long)(frameStart[0]) << "," << (void*)(unsigned long)(frameStart[1]) << "\n";
     }
     frameStart[0] = specialHeader>>8;
     frameStart[1] = (unsigned char)specialHeader;
